+++ packages/cloud-cognitive/src/components/BreadcrumbWithOverflow/BreadcrumbWithOverflow.test.js
@@ -21,6 +21,7 @@ const breadcrumbContent = Array.from(
 
 const breadcrumbItems = breadcrumbContent.map((item) => ({
   href: '/#',
+  id: `id-${item.replace(' ', '_')}`,
   key: item,
   label: item,
   onClick: () => {},
@@ -226,6 +227,21 @@ describe(BreadcrumbWithOverflow.displayName, () => {
     // item 2 contains an overflow menu
     const overflowBtn = screen.getByRole('button');
     expect(overflowBtn).toBeTruthy();
+  });
+
+  it('does not duplicate ids', () => {
+    const plentyOfSpace = (breadcrumbItems.length + 1) * sizes.breadcrumbWidth;
+
+    const { container } = render(
+      <TestBreadcrumbWithOverflow
+        width={plentyOfSpace}
+        overflowAriaLabel="Open and close additional breadcrumb item list."
+        breadcrumbs={breadcrumbItems}
+      />
+    );
+    expect(
+      container.querySelectorAll(`#${breadcrumbItems[0].id}`)
+    ).toHaveLength(1);
   });
 
   it('adds additional properties to an breadcrumb with overflow', () => {