+++ packages/cloud-cognitive/src/components/CreateFullPage/index.js
@@ -7,4 +7,3 @@
 
 export { CreateFullPage } from './CreateFullPage';
 export { CreateFullPageStep } from './CreateFullPageStep';
-export { CreateFullPageSection } from './CreateFullPageSection';