+++ packages/core/story-structure.js
@@ -14,7 +14,21 @@
 // paths but ignored during searching.
 const s = [
   {
-    n: 'Carbon for IBM Products',
+    n: 'Overview',
+    s: [
+      {
+        n: 'Getting started',
+        s: [],
+      },
+      {
+        n: 'Usage examples',
+        s: [],
+      },
+      'o/Examples on CodeSandbox',
+    ],
+  },
+  {
+    n: 'IBM Products',
     s: [
       {
         n: 'Components',
@@ -32,6 +46,7 @@ const s = [
       {
         n: 'Patterns',
         s: [
+          { n: 'AddSelect', s: ['c/AddSelect'] },
           { n: 'About modal', s: ['c/AboutModal'] },
           {
             n: 'Create flows',
@@ -84,8 +99,10 @@ const s = [
           'c/BreadcrumbWithOverflow',
           'c/ButtonMenu',
           'c/ButtonSetWithOverflow',
+          'c/CancelableTextEdit',
           'c/ComboButton',
           'c/ExampleComponent',
+          'c/InlineEdit',
           'c/TearsheetShell',
         ],
       },
@@ -135,10 +152,6 @@ const s = [
         ],
       },
     ],
-  },
-  {
-    n: 'Getting Started',
-    s: [],
   },
 ];